Core/Misc: Really fixed nopch MSVC build
(cherry picked from commit e19bb7a1a5e2ee7881cd9637f46e97982d5f7560)
